refactor: Dto의 id명칭 구체화, 주석 및 import 정리, @Mapping 요청 수정, MindmapDetailResponseDto 생성, ㅡmindmapRepository 수정

+++ src/main/java/com/teamEWSN/gitdeun/mindmap/repository/MindmapRepository.java
@@ -1,8 +1,6 @@
 package com.teamEWSN.gitdeun.mindmap.repository;
 
 import com.teamEWSN.gitdeun.mindmap.entity.Mindmap;
-import com.teamEWSN.gitdeun.mindmap.entity.MindmapType;
-import com.teamEWSN.gitdeun.repo.entity.Repo;
 import com.teamEWSN.gitdeun.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -12,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MindmapRepository extends JpaRepository<Mindmap, Integer> {
+public interface MindmapRepository extends JpaRepository<Mindmap, Long> {
 
     // 사용자가 생성한 확인용 마인드맵 중 가장 최근에 생성된 것(repo 무관)
     @Query("SELECT m FROM Mindmap m " +
